Editor almost fixed.
EDITOR_BUG: when you select ZigZagMover in mover_cls_name and adjust
the value, it gets reset to the default when PropertEditor is opened
second time for the same object. I've got no idea why it happens...
Possibly other Movers also behave this way, although it was not
witnessed. It doesn't seem to ever happen for RechargeBonus.