Editor: new drag'n'drop, info display, item edition features.
The internals of levelview.py have changed drastically. Custom drag'n'drop
is salvaged in favour of the Qt's default implementations, what allows moving
multiple items.
Added mouse position and scene size information displays along with
a statusbar on which they are placed.
New option to clone settings from one item to a group of selected items of
the same type.